+++ defitrack-rest/defitrack-api/src/main/java/io/defitrack/pool/vo/PoolingElementVO.kt
@@ -12,5 +12,5 @@ class PoolingElementVO(
     val protocol: ProtocolVO,
     val dollarValue: Double,
     val id: String,
-    val market: PoolingMarketElementVO
+    val market: PoolingMarketVO
 )
