+++ defitrack-rest/defitrack-api/src/main/java/io/defitrack/staking/vo/StakingPositionVO.kt
@@ -3,9 +3,8 @@ package io.defitrack.staking.vo
 import io.defitrack.network.NetworkVO
 import io.defitrack.protocol.ProtocolVO
 import io.defitrack.token.FungibleToken
-import java.math.BigInteger
 
-data class StakingElementVO(
+data class StakingPositionVO(
     val id: String,
     val network: NetworkVO,
     val protocol: ProtocolVO,
